#64: Phase 2 - @bitbutler/shared types package
Extract IPC contract and shared models into packages/shared so both
Electron and Angular import from one source of truth. App models for
server, electron, torrent-draft, and window re-export from shared.
BitButlerAPI interface now lives in shared/src/ipc.types.ts.
