+++ packages/app/package.json
@@ -1,5 +1,8 @@
 {
   "name": "@bitbutler/app",
   "version": "1.1.0",
-  "private": true
+  "private": true,
+  "dependencies": {
+    "@bitbutler/shared": "*"
+  }
 }
