+++ packages/app/src/app/models/command.model.ts
@@ -1,11 +1,10 @@
+import type { SelectedTorrentInput } from '@bitbutler/shared';
 import { SettingsTabId } from '../pages/settings/settings.interface';
 import { UpdateCheckResponse } from './electron.model';
 import { TorrentDraft } from './torrent-draft.model';
 import { Torrent } from './torrent.model';
 
-export type SelectedTorrentInput =
-  | { name: string; path: string }
-  | { name: string; bytes: number[] };
+export type { SelectedTorrentInput };
 
 export type LimitTargetType = 'global' | 'torrent';
