+++ packages/shared/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@bitbutler/shared",
+  "version": "1.1.0",
+  "private": true,
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "types": "./src/index.ts"
+}
