+++ packages/shared/src/models/window.model.ts
@@ -1,0 +1,7 @@
+export interface WindowState {
+  isMaximized: boolean;
+  isMinimized: boolean;
+  isFullScreen: boolean;
+  width: number;
+  height: number;
+}
